Abolish global constructor and destructor
Why? Because let's be honest dynamic global initization in c++ is just
stupid. There is a reason c does not support dynamic global
initialization. The only way to get around it is through *fancy*
singleton or nifty couter which has its own set of problems.

For singleton, we cannot safely sweep out code(This is not implemented
yet, but would hopefully soon be implemented) for initialization from
memory because the object are constructed for first use and it may well
be that their first use is after we have sweep out the memory used for
the code when we enter userspace.

For nifty counter, we are using extra memory just for those counter
variable which cannot be reclaimed and using nifty counter extensively
would most likely result in ugly and unreadable code as we would have to
define a separate counter and Initializer class everywhere, which are
just boilerplate code to make it work. Furthermore, it does not work
with forward declaration.

One of the added benefit with this new approach is that we can
explicitly leak global objects as we control both their construction and
destruction.

Unless there is added in support for dependency checking between globals
in c++, which is unlikely, this is the final decision.

One may question why is there globals in the first place. Why don't we
just pass around our objects everywhere, and the answer is that it is
simply not feasible, especially if we were to support loadable modules.
A global functional interface, then a massive kernel object that I have
to pass and maintain backward compatibility on.