Move rt::SpinLock into kerneo as core::SpinLock
This is because the only lock that could be reasonably implemented in
freestanding code is a SpinLock, and we all knows that SpinLock is
*BAD*, unless we ask the scheduler to cooperate with use by disabling
preemption, in which case it is less bad, but that is not doable in
freestanding code.

We could have implemented hooks in librt to disable preemption, but then
later, when we want to implement sleeping lock such as mutex, the
interfaces of such hooks become complex.