Refactor kmap/kunmap outside of MemoryMapping class
The fact that kmap and kunmap are a method of the MemoryMapping class is
legacy. In early version of the kernel, we use recursive mapping
technique in order to access physical memory when we are in higher half.
To do so, we must know the *virtual* address of the page tables, which
coincidentally is stored in MemoryMapping strcuture. However,
MemoryMapping strcuture is also used for the management of user memory
area, and thus the design conflates both user and kernel virtual memory
management.  This is starting to be a pain in the ass, especially when
it come to managing kernel initialization order of different subsystem,
which already has a lot of interdependency.

Since user and kernel virtual memory management is now seperated, the
call to core::MemoryMapping::initialize() is hoisted outside of
core::memory::initialize(), which allows us to call it much later during
the initialization process.

As a side effect, the loading of init program hoisted outside of vfs
initialization code into kmain(), since that requires
MemoryMapping::initialize() to have been called.