Move bookkeeping of current task to Scheduler
Previously, this is stored as a static member of core::tasks::Task,
which makes it cumbersome to access or set from the Scheduler. This
is now moved to RunQueue class.

After the change, the method core::tasks::Task::makeCurrent() and
core::tasks::Task::current() is deprecated in favor of
core::tasks::schedule() and core::tasks::current() implemented in
Scheduler.cpp.