Modify rt::logf so that it is atomic
rt::logf used invoke rt::hooks::log multiple time as it parse the
format string and format its argument. When different cpus try to print
at the same time, the output would be mixed together, and become
unreadable. Now, a buffer is used to temporarily store the result of
formatting and rt::hooks::log is invoked only once per rt::logf,
which solve the issue. However, this usage of stack may make the kernel
more prone to stack overflow, which we need to be wary of. Currently,
the size of the buffer is set at 128 bytes.

User of rt::logf is advised try to split the output across multiple
line, so as to not run into the limit on the size of the buffer.