Fix construction of init task and add uses of onResume
Init task is now constructed with state set to RUNNING instead of
RUNNABLE, and is removed via an explicit call to killCurrent.  Calls to
tasks::onResume() is inserted both after the call to schedule() in main
and the end of kernel task instead of call to schedule().  This allows
respective tasks to be removed from the global TasksMap properly by the
reaper task.